requirements.txt
asgiref==3.11.0
channels==4.2.2
daphne==4.1.2
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
google-auth==2.38.0
gspread==6.1.4
psycopg2-binary==2.9.11
requests==2.33.1
sqlparse==0.5.4
tzdata==2025.3
whitenoise==6.9.0
# Python==3.12.10